// src/types/Types.ts
export interface Location {
    lat: number;
    lng: number;
}

export interface Memory {
    id: string;
    story: string;
    location: Location;
    contributorId: string;
    timestamp: number;
    imageUrls: string[];
}

// Global types for Leaflet instances
export type LeafletMap = any;
export type LeafletLayer = any;